More cleanup and another client/server issue
