+++ enderio-armory/src/main/java/com/enderio/armory/client/EnderIOArmoryClient.java
@@ -15,6 +15,7 @@ import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.RegisterItemDecorationsEvent;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.neoforged.neoforge.common.NeoForge;
 
 @EventBusSubscriber(modid = EnderIOArmory.MODULE_MOD_ID, value = Dist.CLIENT, bus = EventBusSubscriber.Bus.MOD)
 @Mod(value = EnderIOArmory.MODULE_MOD_ID, dist = Dist.CLIENT)
@@ -22,6 +23,7 @@ public class EnderIOArmoryClient {
 
     public EnderIOArmoryClient(ModContainer modContainer) {
         modContainer.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
+        NeoForge.EVENT_BUS.addListener(TravelToggleHandler::checkShiftStatus);
     }
 
     @SubscribeEvent
